requirements.txt
flask>=3.0
mysql-connector-python>=8.3
python-dotenv>=1.0
gunicorn>=22.0
boto3>=1.34
